+++ package.json
@@ -32,6 +32,7 @@
     "nodemailer": "^6.6.1",
     "razorpay": "^2.8.0",
     "react": "^16.14.0",
+    "react-accessible-accordion": "^5.0.0",
     "react-anchor-link-smooth-scroll": "^1.0.12",
     "react-bootstrap": "^1.0.0-beta.17",
     "react-countup": "^4.3.3",
