+++ pageComponents/ServiceSection.js
@@ -1,0 +1,50 @@
+import React from "react";
+import {
+  Accordion,
+  AccordionItem,
+  AccordionItemHeading,
+  AccordionItemButton,
+  AccordionItemPanel,
+} from "react-accessible-accordion";
+
+import "react-accessible-accordion/dist/fancy-example.css";
+
+const ServiceSection = () => {
+  return (
+    <div className="service-container">
+      <div className="row">
+        <div className="col-md-6"></div>
+        <div className="col-md-6">
+          <Accordion allowZeroExpanded>
+            <AccordionItem>
+              <AccordionItemHeading>
+                <AccordionItemButton>UX Design</AccordionItemButton>
+              </AccordionItemHeading>
+              <AccordionItemPanel>
+                <p>
+                  Exercitation in fugiat est ut ad ea cupidatat ut in cupidatat
+                  occaecat ut occaecat consequat est minim minim esse tempor
+                  laborum consequat esse adipisicing eu reprehenderit enim.
+                </p>
+              </AccordionItemPanel>
+            </AccordionItem>
+            <AccordionItem>
+              <AccordionItemHeading>
+                <AccordionItemButton>Product Design</AccordionItemButton>
+              </AccordionItemHeading>
+              <AccordionItemPanel>
+                <p>
+                  In ad velit in ex nostrud dolore cupidatat consectetur ea in
+                  ut nostrud velit in irure cillum tempor laboris sed
+                  adipisicing eu esse duis nulla non.
+                </p>
+              </AccordionItemPanel>
+            </AccordionItem>
+          </Accordion>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default ServiceSection;
